requirements.txt
rai-core
rai-whoami
chainlit
